+++ src/cli/default-error-handler.ts
@@ -103,6 +103,7 @@ import RemoteResolverError from '../scope/network/exceptions/remote-resolver-err
 import ExportAnotherOwnerPrivate from '../scope/network/exceptions/export-another-owner-private';
 import ComponentsPendingImport from '../consumer/component-ops/exceptions/components-pending-import';
 import { importPendingMsg } from './commands/public-cmds/status-cmd';
+import { AddingIndividualFiles } from '../consumer/component-ops/add-components/exceptions/addding-individual-files';
 
 const reportIssueToGithubMsg =
   'This error should have never happened. Please report this issue on Github https://github.com/teambit/bit/issues';
@@ -138,6 +139,11 @@ const errorsMap: Array<[Class<Error>, (err: Class<Error>) => string]> = [
   //   err => `error: The compiler "${err.plugin}" is not installed, please use "bit install ${err.plugin}" to install it.`
   // ],
   [FileSourceNotFound, err => `file or directory "${err.path}" was not found`],
+  [
+    AddingIndividualFiles,
+    err => `error: unable to add individual files ("${err.file}"), please add directories only.
+to bypass this error, use --allow-files flag`
+  ],
   [ExtensionFileNotFound, err => `file "${err.path}" was not found`],
   [
     ExtensionNameNotValid,